Moved common functionality to mkspec_common.py
