+++ mkspec/cxx_default.py
@@ -7,6 +7,8 @@ from waflib.Tools.compiler_cxx import cxx_compiler
 import waflib.Tools.gxx as gxx
 from os.path import abspath, expanduser
 import os
+
+import mkspec_common
 """
 Detect and setup the default compiler for the platform
 """
@@ -62,27 +64,28 @@ def configure(conf):
 #         option = None
 #     return option
 
-@conf
-def mkspec_get_toolchain_paths(conf):
-    """
-    :return: the common paths where we may find the g++ binary
-    """
-    # The default path to search
-    path_list = os.environ.get('PATH', '').split(os.pathsep)
+## Moved to mkspec_common.py
+# @conf
+# def mkspec_get_toolchain_paths(conf):
+#     """
+#     :return: the common paths where we may find the g++ binary
+#     """
+#     # The default path to search
+#     path_list = os.environ.get('PATH', '').split(os.pathsep)
 
-    if conf.is_mkspec_platform('mac'):
+#     if conf.is_mkspec_platform('mac'):
 
-        # If the compiler is installed using macports
-        path_list += ['/opt/local/bin']
+#         # If the compiler is installed using macports
+#         path_list += ['/opt/local/bin']
 
-    if conf.is_mkspec_platform('android'):
-        ndk = conf.get_tool_option('android_ndk_dir')
-        ndk = abspath(expanduser(ndk))
-        ndk_path = [ndk, os.path.join(ndk,'bin')]
+#     if conf.is_mkspec_platform('android'):
+#         ndk = conf.get_tool_option('android_ndk_dir')
+#         ndk = abspath(expanduser(ndk))
+#         ndk_path = [ndk, os.path.join(ndk,'bin')]
 
-        return ndk_path
+#         return ndk_path
 
-    return path_list
+#     return path_list
 
 @conf
 def mkspec_get_gxx_binary_name(conf, major, minor):
@@ -126,17 +129,18 @@ def mkspec_check_gxx_version(conf, major, minor):
                    .format(conf.env['CC_VERSION'], major, minor))
 
 
-@conf
-def mkspec_get_ar_binary_name(conf):
-    """
-    :return: The name of the ar binary we are looking for
-             e.g. 'arm-linux-androideabi-ar' for the archiver on android
-    """
+## Moved to mkspec_common.py
+# @conf
+# def mkspec_get_ar_binary_name(conf):
+#     """
+#     :return: The name of the ar binary we are looking for
+#              e.g. 'arm-linux-androideabi-ar' for the archiver on android
+#     """
 
-    if conf.is_mkspec_platform('android'):
-        return 'arm-linux-androideabi-ar'
-    else:
-        return 'ar'
+#     if conf.is_mkspec_platform('android'):
+#         return 'arm-linux-androideabi-ar'
+#     else:
+#         return 'ar'
 
 
 @conf
@@ -261,16 +265,17 @@ def mkspec_gxx_android_configure(conf, major, minor):
     conf.mkspec_gxx_configure(major,minor)
     conf.mkspec_set_android_common()
 
-@conf
-def mkspec_set_android_common(conf):
-    sdk = conf.get_tool_option('android_sdk_dir')
-    sdk = abspath(expanduser(sdk))
-    sdk_path = [sdk, os.path.join(sdk,'platform-tools')]
+## Moved to mkspec_common.py
+# @conf
+# def mkspec_set_android_common(conf):
+#     sdk = conf.get_tool_option('android_sdk_dir')
+#     sdk = abspath(expanduser(sdk))
+#     sdk_path = [sdk, os.path.join(sdk,'platform-tools')]
 
-    conf.find_program('adb', path_list = sdk_path, var='ADB')
+#     conf.find_program('adb', path_list = sdk_path, var='ADB')
 
-    # Set the android define - some libraries rely on this define being present
-    conf.env.DEFINES += ['ANDROID']
+#     # Set the android define - some libraries rely on this define being present
+#     conf.env.DEFINES += ['ANDROID']
 
 
 @conf
